+++ deepinv/tests/test_transform.py
@@ -16,6 +16,7 @@ TRANSFORMS = [
     "shift",
     "rotate",
     "rotate-bilinear",
+    "rotateviashear",
     "scale",
     "reflect",
     "shift+scale",
@@ -108,6 +109,8 @@ def choose_transform(transform_name, device, rng):
         return dinv.transform.Rotate(rng=rng)
     elif transform_name == "rotate-bilinear":
         return dinv.transform.Rotate(rng=rng, interpolation_mode="bilinear")
+    elif transform_name == "rotateviashear":
+        return dinv.transform.RotateViaShear(multiples=180, rng=rng)
     elif transform_name == "rotate3":
         return dinv.transform.Rotate(n_trans=3, rng=rng)
     elif transform_name == "reflect":